+++ modules/core/src/matrix.cpp
@@ -2772,7 +2772,7 @@ static bool ocl_setIdentity( InputOutputArray _m, const Scalar& s )
     k.args(ocl::KernelArg::WriteOnly(m), ocl::KernelArg::Constant(Mat(1, 1, sctype, s)),
            rowsPerWI);
 
-    size_t globalsize[2] = { m.cols, (m.rows + rowsPerWI - 1) / rowsPerWI };
+    size_t globalsize[2] = { static_cast<size_t>(m.cols), static_cast<size_t>((m.rows + rowsPerWI - 1) / rowsPerWI) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -3003,7 +3003,7 @@ static bool ocl_transpose( InputArray _src, OutputArray _dst )
                ocl::KernelArg::WriteOnlyNoSize(dst));
 
     size_t localsize[3]  = { TILE_DIM, BLOCK_ROWS, 1 };
-    size_t globalsize[3] = { src.cols, inplace ? src.rows : divUp(src.rows, TILE_DIM) * BLOCK_ROWS, 1 };
+    size_t globalsize[3] = { static_cast<size_t>(src.cols), static_cast<size_t>(inplace ? src.rows : divUp(src.rows, TILE_DIM) * BLOCK_ROWS), 1 };
 
     return k.run(2, globalsize, localsize, false);
 }
@@ -3485,7 +3485,7 @@ static bool ocl_reduce(InputArray _src, OutputArray _dst,
         kpre.args(ocl::KernelArg::ReadOnly(src),
                   ocl::KernelArg::WriteOnlyNoSize(buf));
 
-        size_t globalSize[2] = { buf_cols, src.rows };
+        size_t globalSize[2] = {  static_cast<size_t>(buf_cols), static_cast<size_t>(src.rows) };
         if (!kpre.run(2, globalSize, NULL, false))
             return false;
