+++ modules/core/src/umatrix.cpp
@@ -694,7 +694,7 @@ void UMat::copyTo(OutputArray _dst, InputArray _mask) const
                    haveDstUninit ? ocl::KernelArg::WriteOnly(dst) :
                                    ocl::KernelArg::ReadWrite(dst));
 
-            size_t globalsize[2] = { cols, rows };
+            size_t globalsize[2] = { static_cast<size_t>(cols), static_cast<size_t>(rows) };
             if (k.run(2, globalsize, NULL, false))
                 return;
         }
@@ -750,7 +750,7 @@ void UMat::convertTo(OutputArray _dst, int _type, double alpha, double beta) con
             else
                 k.args(srcarg, dstarg, alpha, beta, rowsPerWI);
 
-            size_t globalsize[2] = { dst.cols * cn, (dst.rows + rowsPerWI - 1) / rowsPerWI };
+            size_t globalsize[2] = { static_cast<size_t>(dst.cols * cn), static_cast<size_t>((dst.rows + rowsPerWI - 1) / rowsPerWI) };
             if (k.run(2, globalsize, NULL, false))
                 return;
         }
@@ -799,7 +799,7 @@ UMat& UMat::setTo(InputArray _value, InputArray _mask)
                 setK.args(dstarg, scalararg);
             }
 
-            size_t globalsize[] = { cols, (rows + rowsPerWI - 1) / rowsPerWI };
+            size_t globalsize[] = { static_cast<size_t>(cols), static_cast<size_t>((rows + rowsPerWI - 1) / rowsPerWI) };
             if( setK.run(2, globalsize, NULL, false) )
                 return *this;
         }
