+++ modules/imgproc/src/canny.cpp
@@ -113,7 +113,7 @@ static bool ocl_Canny(InputArray _src, OutputArray _dst, float low_thresh, float
     Size esize(size.width + 2, size.height + 2);
 
     UMat mag;
-    size_t globalsize[2] = { size.width, size.height }, localsize[2] = { 16, 16 };
+    size_t globalsize[2] = { static_cast<size_t>(size.width), static_cast<size_t>(size.height) }, localsize[2] = { 16, 16 };
 
     if (aperture_size == 3 && !_src.isSubmatrix())
     {
@@ -230,7 +230,7 @@ static bool ocl_Canny(InputArray _src, OutputArray _dst, float low_thresh, float
                                          size.height, size.width, count);
 
 #define divUp(total, grain) ((total + grain - 1) / grain)
-        size_t localsize2[2] = { 128, 1 }, globalsize2[2] = { std::min(count, 65535) * 128, divUp(count, 65535) };
+        size_t localsize2[2] = { 128, 1 }, globalsize2[2] = { static_cast<size_t>(std::min(count, 65535) * 128), static_cast<size_t>(divUp(count, 65535)) };
 #undef divUp
 
         if (!edgesHysteresisGlobalKernel.run(2, globalsize2, localsize2, false))
