+++ modules/imgproc/src/deriv.cpp
@@ -684,7 +684,7 @@ static bool ocl_Laplacian5(InputArray _src, OutputArray _dst,
     else
         k.args(d2xarg, d2yarg, dstarg, iscale, idelta);
 
-    size_t globalsize[] = { dst.cols * cn / kercn, dst.rows };
+    size_t globalsize[] = { static_cast<size_t>(dst.cols * cn / kercn),  static_cast<size_t>(dst.rows) };
     return k.run(2, globalsize, NULL, false);
 }
