+++ modules/imgproc/src/smooth.cpp
@@ -659,7 +659,7 @@ static bool ocl_boxFilter( InputArray _src, OutputArray _dst, int ddepth,
         wtype = CV_MAKE_TYPE(wdepth, cn), dtype = CV_MAKE_TYPE(ddepth, cn);
 
     const char * const borderMap[] = { "BORDER_CONSTANT", "BORDER_REPLICATE", "BORDER_REFLECT", 0, "BORDER_REFLECT_101" };
-    size_t globalsize[2] = { size.width, size.height };
+    size_t globalsize[2] = { static_cast<size_t>(size.width), static_cast<size_t>(size.height) };
     size_t localsize_general[2] = { 0, 1 }, * localsize = NULL;
 
     UMat src = _src.getUMat();
@@ -2259,7 +2259,7 @@ public:
         #if CV_SSE3
         int CV_DECL_ALIGNED(16) buf[4];
         float CV_DECL_ALIGNED(16) bufSum[4];
-        static const int CV_DECL_ALIGNED(16) bufSignMask[] = { 0x80000000, 0x80000000, 0x80000000, 0x80000000 };
+        static const int CV_DECL_ALIGNED(16) bufSignMask[] = { (int)0x80000000, (int)0x80000000, (int)0x80000000, (int)0x80000000 };
         bool haveSSE3 = checkHardwareSupport(CV_CPU_SSE3);
         #endif
 
@@ -2529,7 +2529,7 @@ static bool ocl_bilateralFilter_8u(InputArray _src, OutputArray _dst, int d,
            ocl::KernelArg::PtrReadOnly(uspace_weight),
            ocl::KernelArg::PtrReadOnly(uspace_ofs));
 
-    size_t globalsize[2] = { dst.cols / sizeDiv, dst.rows };
+    size_t globalsize[2] = { static_cast<size_t>(dst.cols / sizeDiv), static_cast<size_t>(dst.rows) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -2626,7 +2626,7 @@ public:
         #if CV_SSE3
         int CV_DECL_ALIGNED(16) idxBuf[4];
         float CV_DECL_ALIGNED(16) bufSum32[4];
-        static const int CV_DECL_ALIGNED(16) bufSignMask[] = { 0x80000000, 0x80000000, 0x80000000, 0x80000000 };
+        static const int CV_DECL_ALIGNED(16) bufSignMask[] = { (int)0x80000000, (int)0x80000000, (int)0x80000000, (int)0x80000000 };
         bool haveSSE3 = checkHardwareSupport(CV_CPU_SSE3);
         #endif
